update: 共通化するものをcommonへまとめる
